For EZ property sets, added a setter that performs an "atomic" test-and-set: it's atomic in the sense that a single server event will test the variable, and only set it if the previous value was null (i.e. didn't exist). This allows for a level of elementary synchronization between the clients.
The new function on EZGameControl is:
  _gameCtrl.testAndSet (propertyName, newValue[, index])

I hoped to generalize this to test against arbitrary values, but that's 
significantly harder, since properties accept numerous types as values,
and those can have different representations on the client and the server.
So it remains a check against null until we need to generalize it. :)





git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@193 c613c5cb-e716-0410-b11b-feb51c14d237