gameDidEnd(): Clear out _playerOids when a game ends so that all players must re-call playerReady() to restart the game.
playersAllhere();
Start the game if all players call playerReady() when the game is not in play
(previously this only happened if the game was in the pre-game state).

I know of one place this will affect Yohoho which I will fix, but if there's
anywhere else where we are restarting an ended game, fixes will be needed.


git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@329 c613c5cb-e716-0410-b11b-feb51c14d237