how long has this been broken?  In party games, there is a "private" checkbox instead of a "watchable" checkbox, but it was treating it like "watchable", even thought that makes no sense for party games (at present).
As it turns out, party games are broken right now anyway, which is what I'm going to fix next.  
For now, we'll allow party games to use a "Private" checkbox, in case someone wants to play
by themselves.  Private party games won't be given a "join" button.  

This solution is wacky, as is noted in one of the comments.  This will need to be re-visited in the
future, most likely when we do the massive game lobby revamp.


git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@420 c613c5cb-e716-0410-b11b-feb51c14d237