Revamped percentile tracker to support minimum and maximum value tracking.
While testing the new code I discovered that redistribution of large data sets
that approximate a normal curve can cause some funny quantization because the
redistribution algorithm assumes all values in a single bucket are flatly
distributed, but if they actually skew substantially toward the high or low end
of the bucket, then the redistribution generates a choppy curve. Fortunately
this has almost no impact on the percentile curve, which is what's actually
important, it just makes the graph of the contents of the buckets look less
smooth and sexy. Alas.


git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@530 c613c5cb-e716-0410-b11b-feb51c14d237