Allow a custom error handler to be provided to the TableDirector.
If ActionScript supported anonymous classes, TableDirector would not have to
implement InvocationService_ResultListener and could instead define a calldown
handleFailure() method that the client could override by anonymously extending
TableDirector at the point of instantiation and customize the failure handling
thusly. But it doesn't, and so we pollute TableDirector with yet more methods
that many clients probably don't care about.


git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@763 c613c5cb-e716-0410-b11b-feb51c14d237