Make piece prediction harder.
Since the client & server share RNGs, many puzzles (e.g. yohoho's
swordfighting) wind up lending themselves to easy piece prediction; the
board is set up with a seed, and the RNG is only ever used to generate
pieces. Someone snooping that could (and people have) pre-generate all
the pieces in order, letting you snoop ahead.

Now, Board implementations can choose to use certain events to pull bits
from the RNG so that the actual ACTIONS taken perturb things, making
things branch to the point that prediction is prohibitively hard.

A while back, I'd wanted to do this across the board for all actions,
but unfortunately in yohoho we have various non-board-altering events
that come in and are executed in different orders on client and server,
so things fell out of sync pretty quickly. The thought rattled around in
my head again, so I'm doing it again. This time, Boards have to
explicitly opt in on each event, so they can be sure to restrict it to
those that are safe to use.

git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@896 c613c5cb-e716-0410-b11b-feb51c14d237