Fix a bug where the client & server happen to attempt to change location nearly simultaneously.
Rather than ignoring all forced moves from the server, let's instead ignore those which are truly duplicates of our pending move, but merely postpone those which are a separate move until our pending move is complete.

(This works very similarly to the change to LocationDirector in narya)

Relatedly, if a portal traversal is ignored due to staleness (again, due to near-simultaneous client & server moves), tell the client about it so that it doesn't wait around forever thinking it's got a pending move in progress.



git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@970 c613c5cb-e716-0410-b11b-feb51c14d237